Agentic_Workflow/requirements.txt
# LangChain and LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.5.0

# Web Search
tavily-python>=0.5.0

# Data Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.7.0

